wip: tag kit

+++ src/exif/tags.rs
@@ -113,11 +113,7 @@ impl ExifReader {
         let print = if let Some(tag_def) = tag_def {
             if let Some(print_conv_ref) = tag_def.print_conv_ref {
                 // Use new tag kit integration - pass both tag ID and function name
-                registry::apply_print_conv_with_tag_id(
-                    Some(tag_def.id as u32),
-                    print_conv_ref,
-                    &value,
-                )
+                registry::apply_print_conv_with_tag_id(Some(tag_def.id), print_conv_ref, &value)
             } else {
                 value.clone()
             }
